+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+# 5.1.1 - 2026-08-10
+- Added introduction text for TrackingSpamPrevention system setting
+
 # 5.1.0 - 2026-08-03
 - Added an "IP allow list" setting to the UI (General Settings). Existing `iprange_allowlist` config values are migrated to the new `ip_allow_list` system setting and removed from the config file
 - Added an "IP block list" setting to the UI (General Settings) to block tracking requests from specific IP addresses or ranges